+++ test/A4988_Nema_Test.py
@@ -3,44 +3,51 @@
 
 import time 
 import RPi.GPIO as GPIO
+
+# Next 3 lines for development, local library testing import
+# Comment out in production release and change RpiMotorLib.A4988Nema to A4988Nema
 #import sys
 #sys.path.insert(0, '/home/pi/Documents/tech/RpiMotorLib/RpiMotorLib')
+#from RpiMotorLib import A4988Nema
 
-
+# Production installed library import 
 from RpiMotorLib import RpiMotorLib
 
-# 200 steps full revolution
-# 400 half
-# 800 1/4
-# 1600 1/8
-# 3200 1/16
-# __init__(self, direction_pin, step_pin, mode_pins , motor_type):
-# motor_go(clockwise=False, steptype="Full", steps=200, stepdelay=.005, verbose=False, initdelay=.05)
+"""
+# Comment in To Test motor stop put push button to VCC on GPIO 17 
+GPIO.setwarnings(False)
+GPIO.setmode(GPIO.BCM)
+GPIO.setup(17, GPIO.IN, pull_up_down=GPIO.PUD_DOWN)
+"""
 
 def main():
     """main function loop"""
     
-    # ====== tests for motor ====
+    # Comment in To Test motor stop , put push button to VCC on GPIO 17 
+    # GPIO.add_event_detect(17, GPIO.RISING, callback=button_callback)
     
-    #GPIO pins 
+    # ====== Tests for motor ======
+    
     GPIO_pins = (14, 15, 18) # Microstep Resolution MS1-MS3 -> GPIO Pin
     direction= 20       # Direction -> GPIO Pin
     step = 21      # Step -> GPIO Pin
-    
     # Declare an named instance of class pass GPIO-PINs
+    # (self, direction_pin, step_pin, mode_pins , motor_type):
     mymotortest = RpiMotorLib.A4988Nema(direction, step, GPIO_pins, "A4988")
+    
     
     # ====================== section A ===================
     print("TEST SECTION A")
     
+    # motor_go(clockwise, steptype", steps, stepdelay, verbose, initdelay)
     input("TEST: Press <Enter> to continue  Full 180 turn Test1")
-    mymotortest.motor_go(False, "Full" , 100, .005, True, .05)
+    mymotortest.motor_go(False, "Full" , 100, .05, True, .05)
     time.sleep(1)
     input("TEST: Press <Enter> to continue  full 180 clockwise Test2")
-    mymotortest.motor_go(True, "Full" , 100, .005, True, .05)
+    mymotortest.motor_go(True, "Full" , 100, .05, True, .05)
     time.sleep(1)
     input("TEST: Press <Enter> to continue  full 180 no verbose Test3")
-    mymotortest.motor_go(False, "Full" , 100, .005, False, .05)
+    mymotortest.motor_go(False, "Full" , 100, .05, False, .05)
     time.sleep(1) 
     input("TEST: Press <Enter> to continue  timedelay Test4")
     mymotortest.motor_go(True, "Full" , 10, 1, True, .05)
@@ -52,6 +59,7 @@ def main():
     # ========================== section B =========================
     print("TEST SECTION B")
     
+    # motor_go(clockwise, steptype", steps, stepdelay, verbose, initdelay)
     input("TEST: Press <Enter> to continue  half Test1")
     mymotortest.motor_go(False, "Half" , 400, .005, True, .05)
     time.sleep(1)
@@ -65,6 +73,13 @@ def main():
     mymotortest.motor_go(False, "1/16" , 3200, .005, True, .05) 
     time.sleep(1)
     
+"""
+# Comment in for testing motor stop 
+def button_callback(channel): 
+    print("Test file: Stopping motor")
+    mymotortest.motor_stop()   
+"""
+
 # ===================MAIN===============================
 
 if __name__ == '__main__':
@@ -74,7 +89,6 @@ if __name__ == '__main__':
     GPIO.cleanup()
     print("TEST END")
     exit()
-    
     
 
 # =====================END===============================